task(SDK-3110): Addresses the todos to avoid migration on first launch with encLevel 0
